+++ src/plugin.ts
@@ -27,7 +27,7 @@ import {
     HOOKS,
     runPluginHooks
 } from './hooks';
-import { newRxTypeError } from './rx-error';
+import { newRxError, newRxTypeError } from './rx-error';
 
 /**
  * prototypes that can be manipulated with a plugin
@@ -41,6 +41,7 @@ const PROTOTYPES: { [k: string]: any } = {
 };
 
 const ADDED_PLUGINS: Set<RxPlugin | any> = new Set();
+const ADDED_PLUGIN_NAMES: Set<string> = new Set();
 
 /**
  * Add a plugin to the RxDB library.
@@ -53,7 +54,17 @@ export function addRxPlugin(plugin: RxPlugin) {
     if (ADDED_PLUGINS.has(plugin)) {
         return;
     } else {
+
+        // ensure no other plugin with the same name was already added
+        if (ADDED_PLUGIN_NAMES.has(plugin.name)) {
+            throw newRxError('PL3', {
+                name: plugin.name,
+                plugin,
+            });
+        }
+
         ADDED_PLUGINS.add(plugin);
+        ADDED_PLUGIN_NAMES.add(plugin.name);
     }
 
     /**