+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 <!-- CHANGELOG NEWEST -->
 - REFACTOR use faster `areRxDocumentArraysEqual` instead of doing a deep-equal check.
 - ADD check to easier debug wrongly behaving backend during replication.
+- FIX error `dev-mode added multiple times` is removed because it is annoying when using react hot-swap. Instead `addRxPlugin()` will now only throw if a different plugin is added that has the same name as a plugin that was added before.
 <!-- ADD new changes here! -->
 
 <!-- /CHANGELOG NEWEST -->